aksjeradarny-main/requirements-dev.txt
# Development and testing dependencies
pytest==7.4.0
pytest-flask==1.2.0
flask-testing==0.8.1
